Added convenience init and some slider thumbTint color programmatically

+++ Pursuit-Core-iOS-Unit2Final/Pursuit-Core-iOS-Unit2Final/Models/Crayon.swift
@@ -22,14 +22,12 @@ class Crayon {
         self.hex = hex
     }
     
-    
     convenience init(name: String, hex: String) {
-        let redVal = UInt8((String(Array(hex)[0]) + String(Array(hex)[1])), radix: 16)!
-        let greenVal = UInt8((String(Array(hex)[0]) + String(Array(hex)[1])), radix: 16)!
-        let blueVal = UInt8((String(Array(hex)[0]) + String(Array(hex)[1])), radix: 16)!
+        let redVal = UInt((String(Array(hex)[0]) + String(Array(hex)[1])), radix: 16)!
+        let greenVal = UInt((String(Array(hex)[0]) + String(Array(hex)[1])), radix: 16)!
+        let blueVal = UInt((String(Array(hex)[0]) + String(Array(hex)[1])), radix: 16)!
         self.init(name: name, red: Double(redVal), green: Double(greenVal), blue: Double(blueVal), hex: hex)
     }
-    
     
     static let allTheCrayons = [
         Crayon(name: "Almond", red: 239, green: 222, blue: 205, hex: "#EFDECD"),
